% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIX.R
\name{FIX}
\alias{FIX}
\title{SNP statistical test function: FIXED}
\usage{
FIX(z, YFIX, KIN, fn0)
}
\arguments{
\item{z:}{the input one snp genotype matrix for all samples, dim: 1*n, n is the sample counts.The rows represent samples. The columns represent SNPs. If z is NULL, then it will calculate PAR.}

\item{YFIX:}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.}

\item{KIN:}{Kinship matrix. It can be obtained from KIN() function.}

\item{fn0:}{Initial parameters for association test.}
}
\value{
: the test result out is a three-element list:
\enumerate{
\item wald test reuslt: wald test statistic, wald test left tail probability(log), wald test P value;
\item liklihood ratio test(lrt) result: statistics, left tail probability(log), P value;
\item parameters: beta, sigma2, lambda*sigma2, gamma, standand error, lrt statistics, lrt P value,wald test statistic,wald test P value.
}
}
\description{
This function allows you to do statistical test for selected SNP based on FIXED method. It is usually used inside the TEST.SCAN() function.
}
\details{
Additional details...
}
\examples{
FIX(z,YFIX,KIN,fn0)
}
\keyword{cats}
