% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GG.R
\name{GG}
\alias{GG}
\title{Biallelic genotype matrix creation function}
\usage{
GG(G.P)
}
\arguments{
\item{G.P}{Genotype matrix with letter code. Rows represent SNPs, and columns represent samples. Each sample contain one columns (one allele).}
}
\value{
It outputs biallelic letter coded genotype matrix.Rows represent SNPs, and columns represent samples. Each sample contain two columns (two alleles).
}
\description{
This function allows you to convert single allele letter coded matrix to biallelic genotype matrix. It can only be applied for sample which are pure homozyous.
}
\details{
Additional details...
}
\examples{
library(googledrive)
GEN_ID <-  drive_get(as_id("1VPLHa9QWiiey4N5jaUOc516xXo22_fVi"))
drive_download(GEN_ID, overwrite = TRUE)
(load(file="GEN.rda"))
GEN.GG <- GEN[,-(1:2)]
gg<-GG(GEN.GG)
}
\keyword{Biallelic}
\keyword{genotype}
