% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIN.R
\name{KIN}
\alias{KIN}
\title{Kinship matrix calculation function}
\usage{
KIN(gencode)
}
\arguments{
\item{gencode}{Genotype matrix with numeric code. Can be calculated from function GEN.CODE(). It should be a list containing two matrix. The first matrix is additive matrix, for major allele homozygous sample is 1, 0 for heterozygous sample, -1 for minor allele homozygous sample; The second matrix is dominant matrix, 1 for heterozygous sample, 0 for homozygous sample. In each matrix, the rows represent SNPs, and the columns represent samples.}
}
\value{
It outputs kinship matrix(s) in list format.
}
\description{
This function allows you to calculate kinship matrix.
}
\details{
Additional details...
}
\examples{
library(googledrive)
GEN_ID <-  drive_get(as_id("1VPLHa9QWiiey4N5jaUOc516xXo22_fVi"))
drive_download(GEN_ID, overwrite = TRUE)
(load(file="GEN.rda"))

GEN.GG <- GEN[,-(1:2)]
gg<-GG(GEN.GG)
xx<-GEN.CODE(gg)
kin<-KIN(xx)
}
\keyword{kinship}
