% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXED.R
\name{MIXED}
\alias{MIXED}
\title{RANDOM method initial parameter calculation function}
\usage{
MIXED(YFIX, KIN)
}
\arguments{
\item{YFIX}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.}

\item{KIN}{Kinship matrix. It can be obtained from KIN() function.}
}
\value{
: Initial parameter fn0 which can be used in RANDOM function.
}
\description{
This function allows you to calculate the initial parameter for RANDOM function. It is usually used inside the TEST.SCAN() function.
}
\details{
Additional details...
}
\examples{
MIXED(YFIX,KIN)
}
