% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProF.R
\name{ProF}
\alias{ProF}
\title{Allele frequency calculation function}
\usage{
ProF(GG)
}
\arguments{
\item{GG}{Genotype matrix with letter code. Rows represent SNPs, and columns represent samples. Each sample contain two columns (two alleles).}
}
\value{
It outputs a data frame with three columns: minor allele frequency, major allele, minor allele.
}
\description{
This function allows you to calculate allele frequency.
}
\details{
Additional details...
}
\examples{
library(googledrive)
GEN_ID <-  drive_get(as_id("1VPLHa9QWiiey4N5jaUOc516xXo22_fVi"))
drive_download(GEN_ID, overwrite = TRUE)
(load(file="GEN.rda"))

GEN.GG <- GEN[,-(1:2)]
gg<-GG(GEN.GG)
prof <- ProF(gg)
}
\keyword{allele}
\keyword{frequency}
