% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANDOM.R
\name{RANDOM}
\alias{RANDOM}
\title{SNP statistical test function: RANDOM}
\usage{
RANDOM(z, YFIX, KIN, Theta)
}
\arguments{
\item{z:}{the input one snp genotype matrix for all samples, dim: 1*n, n is the sample counts.The rows represent samples. The columns represent SNPs.}

\item{YFIX:}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.}

\item{KIN:}{Kinship matrix. It can be obtained from KIN() function.}

\item{theta:}{Initial parameters for association test.}
}
\value{
: the test result out is a three-element list:
\enumerate{
\item wald test reuslt: wald test statistic, wald test left tail probability(log), wald test P value;
\item liklihood ratio test(lrt) result: statistics, left tail probability(log), P value;
\item parameters: beta, sigma2, lambda*sigma2, gamma (two elements for heterozygote), standand error(four elements for heterozygote), wald test statistic,wald test P value,lrt statistics, lrt P value,.
}
}
\description{
This function allows you to do statistical test for selected SNP based on RANDOM method. It is usually used inside the TEST.SCAN() function.
}
\details{
Additional details...
}
\examples{
RANDOM(z,YFIX,KIN,Theta)
}
\keyword{cats}
