% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEL.HAP.R
\name{SEL.HAP}
\alias{SEL.HAP}
\title{Haplotype association detection function}
\usage{
SEL.HAP(
  MAP.S = NULL,
  POS.S = NULL,
  GEN,
  YFIX,
  KIN,
  nHap = 2,
  method,
  p.threshold,
  RR0 = NULL,
  TEST,
  PAR = NULL
)
}
\arguments{
\item{MAP.S:}{SNP map matrix containing the genomic positions for SNPs the user wants to scan. It should contain 2 columns: chromosome and base pair positions respectively. It can be NULL.}

\item{POS.S:}{SNP map index (relevant to the full SNP map file) referring the genomic position information for SNPs which the user want to scan, It should be a vector of numbers. It can be NULL.}

\item{GEN:}{Letter code genotype matrix with SNP genomic position information. The rows represent SNPs. The columns represent samples and each sample has two columns. The first two columns are chromosome and base pair positions.}

\item{YFIX:}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.Note: the input data must be the matrix class.}

\item{KIN:}{Kinship matrix. It can be obtained from KIN() function.}

\item{nHap:}{Initial haplotype length user want to start with. The default value is 2.}

\item{method:}{Association model user want to use. It could be FIXED or RANDOM.}

\item{p.threshold:}{The p value threshold for initial significant haplotype selection. It should be a loose threshold to keep more seed haplotype for extension step.}

\item{RR0:}{Initial significant haplotype selection result. It can be NULL.}

\item{TEST:}{The statistical test user want to use for significance test. 1: wald test; 2: lrt test, c(1,2) for both.}

\item{PAR:}{Initial parameters for association test. The default is NULL. It can be calculated through function TEST.SCAN().}
}
\value{
SEL.HAP function output a list containing two element. The first element is FINAL haplotype scanning and statistical test results. The second selement is initial haplotype statistical test results.
Final matrix has 9 columns: chrosome, position index of haplotype start SNP, position index haplotype end SNP, position of haplotype start SNP, position haplotype end SNP, statistic,  left tail probability(log), P value, numbers of different haplotypes detected;
}
\description{
This function allows you to detect target phenotype associated haplotypes.
}
\details{
Additional details...
}
\examples{
SEL.HAP(MAP.S=NULL, POS.S=NULL, GEN, YFIX, KIN, nHap=2, method, p.threshold, RR0=NULL, TEST, PAR=NULL)
}
\keyword{cats}
