% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUDO.GG.R
\name{SUDO.GG}
\alias{SUDO.GG}
\title{SUDO letter coded genotype matrix creation function}
\usage{
SUDO.GG(gencode.a)
}
\arguments{
\item{gencode.a:}{Genotype matrix with numeric code: -1/1. The rows represent SNPs. The columns represent samples.}
}
\value{
It outputs sudo letter coded genotype matrix in list format.
}
\description{
This function allows you to convert numeric coded genptype matrix into a sudo letter coded genotype matrix. It can only be applied for sample which are pure homozyous.
}
\details{
Additional details...
}
\examples{
SUDO.GG(gencode.a)
}
