% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.HAP.R
\name{test.HAP}
\alias{test.HAP}
\title{Haplotype association statistical test function}
\usage{
test.HAP(HAP.X, YFIX, KIN, method, PAR)
}
\arguments{
\item{HAP.X:}{letter code geno matrix for the position defined. The rows represent samples and each sample has two rows. The columns represent SNPs.}

\item{YFIX:}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.Note: the input data must be the matrix class.}

\item{KIN:}{Kinship matrix. It can be obtained from KIN() function.}

\item{method:}{Association model user want to use. It could be FIXED or RANDOM.}

\item{PAR:}{Initial parameters for association test. It can be calculated through function TEST.SCAN().}
}
\value{
: SEL.HAP function output # a list with three elements:
\enumerate{
\item a list with two elements:
}
\enumerate{
\item wald test reuslt: wald test statistic, wald test left tail probability(log), wald test P value, numbers of different haplotypes detected;
\item liklihood ratio test(lrt) result: statistics, left tail probability(log), P value,numbers of different haplotypes detected;
}
\enumerate{
\item z matrix: code of haplotypes;
\item identified haplotypes (numerical code).
}
}
\description{
This function allows you to do statistical test for selected haplotypes.It is usually used inside the SEL.HAP() function.
}
\details{
Additional details...
}
\examples{
test.HAP(HAP.X,YFIX,KIN,method,PAR)
}
\keyword{cats}
