% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GG.R
\name{GG}
\alias{GG}
\title{Biallelic genotype matrix creation function}
\usage{
GG(G.P)
}
\arguments{
\item{G.P}{Genotype matrix with letter code. Rows represent SNPs, and columns represent samples. Each sample contain one columns (one allele).}
}
\value{
It outputs biallelic letter coded genotype matrix.Rows represent SNPs, and columns represent samples. Each sample contain two columns (two alleles).
}
\description{
This function allows you to convert single allele letter coded matrix to biallelic genotype matrix. It can only be applied for sample which are pure homozyous.
}
\details{
Additional details...
}
\keyword{Biallelic}
\keyword{genotype}
