% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProF.R
\name{ProF}
\alias{ProF}
\title{Allele frequency calculation function}
\usage{
ProF(GG)
}
\arguments{
\item{GG}{Genotype matrix with letter code. Rows represent SNPs, and columns represent samples. Each sample contain two columns (two alleles).}
}
\value{
It outputs a data frame with three columns: minor allele frequency, major allele, minor allele.
}
\description{
This function allows you to calculate allele frequency.
}
\details{
Additional details...
}
\keyword{allele}
\keyword{frequency}
