% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEL.SNP.R
\name{SEL.SNP}
\alias{SEL.SNP}
\title{SNP association detection function}
\usage{
SEL.SNP(GEN, YFIX, KIN, method, PAR = NULL)
}
\arguments{
\item{GEN}{Letter code genotype matrix with SNP genomic position information. The rows represent SNPs. The columns represent samples and each sample has two columns. The first two columns are chromosome and base pair positions.}

\item{YFIX}{Phenotype input matrix. The first column is target phenotype data. The rest columns are FIXED traits user want to put into the model. If no FIXED traits, put 1 in the second column.Note: the input data must be the matrix class.}

\item{KIN}{Kinship matrix. It can be obtained from KIN() function.}

\item{method}{Association model user want to use. It could be FIXED or RANDOM.}

\item{PAR}{Initial parameters for association test. The default is NULL. It can be calculated through function TEST.SCAN().}
}
\value{
SEL.HAP function output a list containing three element. 1. Wald test result (chr, pos, wald test statistic, wald test left tail probability (log), wald test P value; 2.liklihood ratio test(lrt) result (chr, pos,statistics, left tail probability(log), P value)); 3. full result with 9 elements (beta, sigma2, lambda*sigma2, gamma, standand error, wald test statistic,wald test P value,lrt statistics, lrt P value);
}
\description{
This function allows you to detect target phenotype associated SNPs
}
\details{
Additional details...
}
\examples{
library(googledrive)
GEN_ID <-  drive_get(as_id("1VPLHa9QWiiey4N5jaUOc516xXo22_fVi"))
drive_download(GEN_ID, overwrite = TRUE)
(load(file="GEN.rda"))

Y_ID <-  drive_get(as_id("1AF3XGQr-MwsR928NRLM5X6SwYx20NcUA"))
drive_download(Y_ID, overwrite = TRUE)
(load(file="Y.rda"))

GEN.GG <- GEN[,-(1:2)]
gg<-GG(GEN.GG)
xx<-GEN.CODE(gg)
kin<-KIN(xx)
CP<-matrix(as.numeric(GEN[,1:2]),nrow(GEN),2)
gen<-cbind(CP,gg)
method<-"RANDOM"
YFIX <- as.matrix(Y[,2:3])
PAR<-TEST.SCAN(YFIX,NULL,KIN=kin,method,NULL)
snp_scan <- SEL.SNP(gen, YFIX, KIN=kin, method, PAR=PAR)
}
\keyword{GWAS}
\keyword{SNP-based}
