% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis}
\alias{vis}
\title{GWAS visualization function: vis}
\usage{
vis(
  T.Name,
  snp_file,
  hapi_file,
  hap_file,
  delim = "\\t",
  sig_line = 5e-08,
  ylim,
  snp_sig_size = 8,
  snp_sig_alpha = 0.8,
  snp_nosig_size = 1,
  snp_no_sig_alpha = 0.3,
  hapi_sig_lineend = "round",
  hapi_sig_linewidth = 5,
  hapi_sig_alpha = 1,
  hapi_nosig_lineend = "round",
  hapi_nosig_linewidth = 1,
  hapi_nosig_alpha = 0.3,
  hap_sig_lineend = "round",
  hap_sig_linewidth = 3,
  hap_sig_alpha = 1,
  hap_nosig_lineend = "round",
  hap_nosig_linewidth = 1,
  hap_nosig_alpha = 0.3,
  snp_color = "#93b5cf",
  hapi_color = "#ffa60f",
  hap_color = "#f03752"
)
}
\arguments{
\item{T.Name}{Phenotype name, shown in the plot name;}

\item{snp_file}{Path of file contain SNP-based GWAS result. The file contain four column: chromosome, base pair position, p value and SNP ID;}

\item{hapi_file}{Path of file contain initial di-SNP GWAS result. The file contain five column: chromosome, start base pair position, end base pair position,p value and SNP ID;}

\item{hap_file}{Path of file contain haplotype-based GWAS final result. The file contain four column: chromosome, start base pair position, end base pair position, p value and SNP ID;}

\item{delim}{Delimiter of input snp, hapi and hap file, default is tab;}

\item{sig_line}{Significant p value threshold;}

\item{ylim}{Maximum number of y axis;}

\item{snp_sig_size}{Point size of significant SNP signals, default is 8;}

\item{snp_sig_alpha}{Point transparence of significant SNP signals, default is 0.8;}

\item{snp_nosig_size}{Point size of non-significant SNP signalsdefault is 1;}

\item{snp_no_sig_alpha}{Point transparence of non-significant SNP signals, default is 0.3;}

\item{hapi_sig_lineend}{Line end style (round, butt, square) of initial haplotype significant signals, default is "round";}

\item{hapi_sig_linewidth}{Line width of initial haplotype significant signals, default is 5;}

\item{hapi_sig_alpha}{Line transparence of initial haplotype significant signals, default is 1;}

\item{hapi_nosig_lineend}{Line end style (round, butt, square) of initial haplotype non-significant signals, default is "round";}

\item{hapi_nosig_linewidth}{Line width of initial haplotype non-significant signals, default is 1;}

\item{hapi_nosig_alpha}{Line transparence of initial haplotype non-significant signals, default is 0.3;}

\item{hap_sig_lineend}{Line end style (round, butt, square) of haplotype significant signals, default is "round";}

\item{hap_sig_linewidth}{Line width of haplotype significant signals, default is 3;}

\item{hap_sig_alpha}{Line transparence of haplotype significant signals, default is 1;}

\item{hap_nosig_lineend}{Line end style (round, butt, square) of haplotype non-significant signals, default is "round";}

\item{hap_nosig_linewidth}{Line width of haplotype non-significant signals, , default is 1;}

\item{hap_nosig_alpha}{Line transparence of haplotype non-significant signals, default is 0.3;}

\item{snp_color}{SNP signal color, default is "#93b5cf";}

\item{hapi_color}{Initial haplotype signal color, default is "#ffa60f";}

\item{hap_color}{Haplotype signal color, default is "#f03752";}
}
\value{
manhattan plot containing SNP-based GWAS, initial haplotype scan and Haplotype-based GWAS result.
}
\description{
This function allows you to make manhattan plot containing SNP-based GWAS, initial haplotype scan and Haplotype-based GWAS result.
}
\details{
Input file p value column should not contain missing value and 0. If the original p value is 0 (which is too small to calculate by the software), please compute p value manually or set a standard small p value for plot the figure.
}
\keyword{Manhattan}
\keyword{plot}
